alpago073/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-multiselectfield==0.1.13
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
